+++ src/class/styles/sizing/max_height.rs
@@ -11,24 +11,24 @@ pub fn max_h(px: f32) -> MaxHeight {
     MaxHeight(Val::Px(px))
 }
 
-pub fn max_h_vh(percent: f32) -> MaxHeight {
-    MaxHeight(Val::Vh(percent))
+pub fn max_h_vh(fraction: f32) -> MaxHeight {
+    MaxHeight(Val::Vh(fraction * 100.))
 }
 
-pub fn max_h_vmin(percent: f32) -> MaxHeight {
-    MaxHeight(Val::VMin(percent))
+pub fn max_h_vmin(fraction: f32) -> MaxHeight {
+    MaxHeight(Val::VMin(fraction * 100.))
 }
 
-pub fn max_h_vmax(percent: f32) -> MaxHeight {
-    MaxHeight(Val::VMax(percent))
+pub fn max_h_vmax(fraction: f32) -> MaxHeight {
+    MaxHeight(Val::VMax(fraction * 100.))
 }
 
 pub fn max_h_div(n: u32, d: u32) -> MaxHeight {
     MaxHeight(Val::Percent((n as f32 / d as f32) * 100.0))
 }
 
-pub fn max_h_fract(percent: f32) -> MaxHeight {
-    MaxHeight(Val::Percent(percent))
+pub fn max_h_fract(fract: f32) -> MaxHeight {
+    MaxHeight(Val::Percent(fract * 100.))
 }
 
 #[derive(Debug, Clone, Copy, PartialEq)]